refactor: backend to facilitate testing
